First attempt at a neural network, and GPU support
Attempted a U-Net architecture for a model that learns to retrieve the original gradients from a final Perlin noise image (on a sphere). Largely a failure as loss does not appear to decrease, but committing the progress so far since the general network structure seems like a good starting point.

The first part of the network receives and outputs images of the same shape, representing a sphere in cylindrical projection. The input has one channel while the output has many: one per layer of noise. This is the part that uses a U-Net structure. Next, the output images are each mapped into 6 separate 3D vector fields, representing the faces of the sphere when projected to a cube. The dimensions of each face match the dimensions of each layer of noise, so this is the intended output of the network.

For training, this output is passed to the function that generates Perlin noise from gradients. The loss function is then the MSE between this noise and the original input noise.

Although GPU support was added, it seems to perform worse than CPU. This is probably due to the last part of the forward function which iterates over lists. In future this should be fully vectorized into Tensors, hopefully improving performance.