First working model
Network is finally producing results that appear valid, with outputs becoming noticeably similar to inputs! Loss has decreased to 0.005 and is possibly still dropping.

The inputs are now pre-processed by creating multiple shifted versions of the input image, in which each corner of a Perlin grid cell become superimposed onto the center of that cell. Each shifted image becomes a separate channel in the input. The reasoning behind this is that each pixel value is the result of a Perlin interpolation between the nearby corners, so providing these additional channels moves information spatially closer to the regions where that information is important. Additionally, this solves the problem of faces requiring some information from nearby faces, which was previously (unsuccessfully) approached by padding the face cropping window.